+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added the `category` attribute to the jurisdictions ([PR 2104](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2104))
+
 ## [6.2.0] - 2025-09-22
 
 ### Added