+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix issue where organisation sector name was not being shown correctly ([PR 2126](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2126))
+- Fix issue where the date validation for checking a real date was not triggering ([PR 2134](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2134))
 
 ## [6.2.0] - 2025-09-22